Moved Main component into its own folder
