+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Layout from '../components/Layout/Layout'
 
 import Header from '../components/Header/Header'
-import Main from '../components/Main'
+import Main from '../components/Main/Main'
 import Footer from '../components/Footer/Footer'
 
 class IndexPage extends React.Component {
